services/paddleocr/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
paddleocr==2.7.3
paddlepaddle==2.5.2
opencv-python==4.8.1.78
Pillow==10.0.1
pdf2image==1.16.3
numpy==1.24.3